% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addToZero}
\alias{addToZero}
\title{Add X to Zero Values}
\usage{
addToZero(x, amount = 1L)
}
\arguments{
\item{x}{A numerical value; \code{NA}s and \code{Inf}s -- including \code{-Inf}s.}

\item{amount}{numerical value (doubles, floats, and integers) to add to the
zero (0) values.}
}
\value{
numerical value
}
\description{
Sometimes you need to increase a zero value (0) to prevent \code{-Inf} and
\code{Inf} values due to an accidental divide by zero (0).
}
\examples{
addToZero(x=0, amount=1L)
# [1] 1
addToZero(x=0, amount=0.001)
# [1] 0.001


}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
