% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_datasets.R
\name{classification.check}
\alias{classification.check}
\title{Classification Check}
\usage{
classification.check(data, col.oi, ct, ct.col)
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame} of interest}

\item{col.oi}{string indicating the column with classes of interest}

\item{ct}{\code{tibble} or \code{data.frame} conversion table with a column with the
information to be converted and the information to be converted to}

\item{ct.col}{string indicating the column to match to the classes in the \code{col.oi}}
}
\value{
Indicator of missing classifications. See Examples.
}
\description{
Identify when a new class in the list of \strong{long} responses that
need to be converted to their \strong{short}ened forms.
}
\examples{
\dontrun{
 classification.check(data=data.oi, col.oi="Animal", ct=ToFrom.ct, ct.col="long")
 # No new classes in the dataset of interest.

 classification.check(data=data.oi, col.oi="Animal", ct=ToFrom.ct, ct.col="long")
 # There are unmatched values in the data of interest:
 #  -> Dogs
 #  -> Cats
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
