% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_text.R
\name{clean.cntrl}
\alias{clean.cntrl}
\title{Convert Control Characters to Spaces}
\usage{
clean.cntrl(text)
}
\arguments{
\item{text}{A string of words forming a sentence or phrase.}
}
\value{
string of cleaned characters forming a sentence or phrase.
}
\description{
Converts control characters into physical spaces.
}
\details{
One of several text cleaning functions. The function replaces the
following control characters \verb{\\n}, \verb{\\r}, \verb{\\t}, \verb{\\v}, and \verb{\\f} (those
identified by the \verb{[:cntrl:]} regular expression) with a series of three (3)
physical spaces. While multiple, concurrent and connected control characters
(\verb{\\n\\r\\t\\v\\f}) would be replace by three spaces, control characters separated
by any printable character (defined by \verb{[:print:]}), even spaces, are retained;
see example.
}
\examples{
comment <- "All day I see class mates and send e-mails.\n\r\t\\v\fI miss East Lansing."
clean.cntrl(text=comment)
# [1] "All day I see class mates and send e-mails.   I miss East Lansing."

comment <- "All day I see class mates and send e-mails.\n  \r  \t  \\v\fI miss East Lansing."
clean.cntrl(text=comment)
# [1] "All day I see class mates and send e-mails.                  I miss East Lansing."

}
\seealso{
Other "text cleaning": 
\code{\link{convert.spaces}()},
\code{\link{extract.emails}()}
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\concept{"text cleaning"}
