% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{clean.comment}
\alias{clean.comment}
\title{Clean Comments}
\usage{
clean.comment(comment)
}
\arguments{
\item{comment}{string of characters (usually text) to be evaluated and unified.
Alternatively, the column containing the comments can be passed to the
function as often is done via \code{\link[dplyr:mutate]{dplyr::mutate()}}}
}
\value{
vector of comments where predetermined terms are cleaned.
\emph{\strong{NOTE}}: Some cleaned terms are returned in camel case.
For example, "e-mail" is returned as "eMail" and "face to face" and
"face-to-face" are returned as "faceTOface".
}
\description{
Normalize (aka \emph{clean}) comments to reduce the number of small
counts for similar or related words.
}
\details{
Often similar words or ideas have different forms depending on their
location within a sentence. For example, the term "advisor" can have three
different forms within a comment: "adviser", "advisers", "advisors" and all
three are converted (transformed) to "advisor".
}
\examples{
comment <- "All day I see class mates and send e-mails.   I miss East Lansing."

clean.comment(comment=comment)
# "allDay i see classmate and send eMail. i miss eastLansing."

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
