% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.DATE}
\alias{clean.DATE}
\title{Clean Date}
\usage{
clean.DATE(dates)
}
\arguments{
\item{dates}{string with date of birth-like data information.}
}
\value{
date
}
\description{
This function expands \code{\link[=clean.DoB]{clean.DoB()}} function and uses the
\code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}} function to determine the provided date by trying
\itemize{
\item Month - Day - Year
\item Year - Month - Day
\item Day - Month - Year
\item Month - Year
\item Year -Month
}

The function determines if the provided date contains month-day-year or month-
year information.

\strong{This function is currently not available. Need to add a logical parameter.}
A date with a year five-years before the current year or later
is reduced by 100-years. For example, if the resulting year is 2020 (this
description was written in 2022) the returned date has a year of 1920.
}
\examples{
clean.DATE("03301995")
# "1995-03-30"

clean.DATE("03/30/1995")
# "1995-03-30"

clean.DATE("03/1995")
# "1995-03-30"

clean.DATE("031995")
# "1995-03-30"

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
