% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.DoB}
\alias{clean.DoB}
\title{Clean Date of Birth}
\usage{
clean.DoB(DoB)
}
\arguments{
\item{DoB}{string with date of birth-like data information}
}
\value{

}
\description{
Clean Date of Birth
}
\examples{
clean.DoB("03301995")
# "1995-03-30"

clean.DoB("03/30/1995")
# "1995-03-30"

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
