% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.OStype}
\alias{clean.OStype}
\title{Get Operating System Name}
\usage{
clean.OStype(OperatingSystem)
}
\arguments{
\item{OperatingSystem}{A string containing the operating system obtained by
a Qualtrics survey; \emph{e.g.}, "Windows NT 4.0 SP6a"}
}
\value{
string with the \emph{generic} operating system name; \emph{e.g.}, "Windows"
}
\description{
Return the \emph{generic} name of the operating system.
}
\examples{
clean.OStype("Windows NT 10.0")
# "Windows"

clean.OStype("Windows NT 4.0 SP6a")
# "Windows"

clean.OStype("CrOS x86_64 13505.63.0")
# "CrOS"

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
