% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.OSversion}
\alias{clean.OSversion}
\title{Get Main Version Number}
\usage{
clean.OSversion(Version.number)
}
\arguments{
\item{Version.number}{A string with the version number; \emph{e.g.}, "10.15.7".
Because version numbers can have multiple points, they are not considered
numbers within \code{R} and thus they are strings.}
}
\value{
integer of the main version number
}
\description{
Returns the main version number.
}
\examples{
clean.OSversion("10.15.7")
# 10

clean.OSversion("11.3")
# 11

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
