% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.YoB}
\alias{clean.YoB}
\title{Clean Year of Birth}
\usage{
clean.YoB(YoB)
}
\arguments{
\item{YoB}{string with date of birth-like data information}
}
\value{
integer with the determined year of birth
}
\description{
Clean Year of Birth
}
\examples{
clean.YoB("1969")
# 1969

clean.YoB("01/23/1969")
# 1969

clean.YoB("01/23/69")
# 1969

clean.YoB("23 January 1969")
# 1969

clean.YoB("19690123")
# 1969

clean.YoB("69")
# 1969

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
