% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_ttest.R
\name{combinations.t.test}
\alias{combinations.t.test}
\title{t-test Combinations}
\usage{
combinations.t.test(
  data,
  alternative = "two.sided",
  paired = FALSE,
  conf.level = 0.99,
  na.rm = TRUE,
  CER = 0.2
)
}
\arguments{
\item{data}{The \code{list} with the values to be compared. Each list should have
a name otherwise names will be autogenerated.}

\item{alternative}{The type of t-test analysis; default: \code{"two.sided"}}

\item{paired}{\code{logical} indicating if the data is the result of a paired
sample; default: \code{FALSE}}

\item{conf.level}{The confidence level; default: \code{0.99}}

\item{na.rm}{\code{logical} indicating if \code{NA} values are removed; default: \code{TRUE}}

\item{CER}{\code{numerical} value indicating the control event rate (CER); default: \code{0.20}}
}
\value{
A \code{data.frame} or \code{tibble} with the \code{t-value}, degrees of freedom
(\code{df}), and \code{p-value}.
}
\description{
Calculate all t-test Combinations
}
\details{
Calculate all t-test combinations within a list of values.
}
\examples{
\dontrun{
  values.ls <- list(a=1:3, b=1:5, c=1:10, d=1:20, e=letters[1:20])
  values.tb <- tibble::tibble(a=1:10, b=11:20, c=21:30, d=31:40, e=letters[1:10])
  combinations.t.test(values.ls)
  combinations.t.test(values.tb)
}

}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
(\url{https://github.com/emilioxavier})
}
