% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{comment.summary}
\alias{comment.summary}
\title{Comment Summary}
\usage{
comment.summary(data, comment.col)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with the comments of interest}

\item{comment.col}{column name with the comments; \emph{e.g.}, \code{"comments"}}
}
\value{
augmented \code{tibble} (or \code{data.frame}) with the following additional
information:
\itemize{
\item is the comment a non-comment (see \code{\link[=is.none]{is.none()}})
\item number of words in the comment (see \code{\link[=word.count]{word.count()}})
\item number of keyword areas contained within the comment
}

A summary for all comments is returned to the user.
}
\description{
Brief, numeric summary of the comments added to the \code{tibble}
(or \code{data.frame}).
}
\examples{
\dontrun{
comment.summary(data=comment.data, comment.col="comment.clean")
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
