% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compare.dataset}
\alias{compare.dataset}
\title{Compare Datasets}
\usage{
compare.dataset(data)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) to be compared}
}
\value{
three matrices are returned.
\itemize{
\item \strong{Tanimoto.Indices}: pairwise Tanimoto coefficients for all samples
\item \strong{Overlap.Counts}: number of overlapping samples for each area of interest.
Must be read column-wise because the number of counts is for each row
area within the column.
\item \strong{Overlap.Ratios}: ratio (aka fraction) of number of overlapping samples
for each area of interest. Must be read column-wise because the ratio
is for each row area within the column.
}
}
\description{
Calculates the amount of similarity (or dissimilarity based on
how you are looking at the data) within a dataset.
}
\details{
The number and ratio (aka fraction) of overlapping samples within
each area of interest along with the Tanimoto coefficient/index are
calculated. The returned Tanimoto coefficients/indice matrices are symmetrical
around the diagonal, but the count and ratio matrices are not. \emph{\strong{They can
be interpreted two different ways.}}

\strong{Count and Ratio Matrices}
\itemize{
\item The diagonal of the count and ratio matrices contain the number of
samples containing the area of interest.
\item Within each column, the returned counts or ratio values indicate the number (ratio)
of samples within each area that contains both areas (the column and the
row). The returned column values are always to be compared to the column
of interest. The returned count will \emph{\strong{never}} be more than the diagonal
value of the column.
\item Each row, especially for the ratio matrices, should be interpreted as
the similarity of each column to each other. For example, the ratios within
a row
\item Within each row, the returned counts or ratio values indicate the number (ratio)
of samples within each area that contains both areas (the column and the
row).
}
}
\examples{
\dontrun{
compare.dataset(data)
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
