% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.ACT2SAT}
\alias{convert.ACT2SAT}
\title{Convert ACT Score to SAT Score}
\usage{
convert.ACT2SAT(ACT.score)
}
\arguments{
\item{ACT.score}{ACT score as a number (float or integer). \code{NA}s and text (aka
\code{as.character()}) are converted to integer values via \code{as.integer()}.}
}
\value{
integer of the corresponding SAT score
}
\description{
Convert a user provided ACT score to the corresponding SAT score.
This function uses the \href{https://www.princetonreview.com/college-advice/act-to-sat-conversion}{Princeton Review}
conversion table. Because a range of SAT values equal a single ACT score, the
mean of the SAT scores provided in
\href{https://www.princetonreview.com/college-advice/act-to-sat-conversion}{The Princeton Review's ACT to SAT Score Conversion Chart}
is used for this conversion. Please see the the \link{ACT.2.SAT}
dataset for details.
}
\examples{
ACT.score <- c(25, "34", NA, 25.25, "NA", 50)
convert.ACT2SAT(ACT.score)
# [1] 1215 1535   NA 1215   NA   NA

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
