% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{convert.fps2binary}
\alias{convert.fps2binary}
\title{Convert Integer Fingerprints into Binary Representation}
\usage{
convert.fps2binary(data, id.col)
}
\arguments{
\item{data}{data.frame or tibble with fingerprint values}

\item{id.col}{string indicating the identifying column}
}
\value{
tibble of fingerprints with logical (binary) values
}
\description{
Convert integer fingerprints into binary (logical) representation.
}
\examples{
\dontrun{
  library(tidyverse)
  where2eat <- tibble::tibble(what=c("BBQ", "Burgers", "Pizza"),
                              loc=c(1,2,3),
                              size=c(4,2,6),
                              cost=c(10,9,8))
  make.fps(data=where2eat, id.col="what", cols.oi=c("loc", "size", "cost")) |>
    convert.fps2binary()
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
