% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.SAT2ACT}
\alias{convert.SAT2ACT}
\title{Convert SAT Score to ACT Score}
\usage{
convert.SAT2ACT(SAT.score)
}
\arguments{
\item{SAT.score}{SAT score as a number (float or integer)}
}
\value{
integer of the corresponding ACT score
}
\description{
Convert a user provided SAT score to the corresponding ACT score.
This function uses the \href{https://www.princetonreview.com/college-advice/act-to-sat-conversion}{Princeton Review}
conversion table. Because a range of SAT values equal a single ACT score, a
range of SAT scores will return the same ACT score. Please see
\href{https://www.princetonreview.com/college-advice/act-to-sat-conversion}{The Princeton Review's ACT to SAT Score Conversion Chart}
or the \link{SAT.2.ACT} dataset for details.
}
\examples{
SAT.score <- c(1326, "1444", NA, 1444.44, "NA", 3600)
convert.SAT2ACT(SAT.score)
# [1] 28 31 NA 31 NA NA

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
