% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_text.R
\name{convert.spaces}
\alias{convert.spaces}
\title{Convert all Spaces to "Normal" Spaces}
\usage{
convert.spaces(text)
}
\arguments{
\item{text}{A string of words forming a sentence or phrase.}
}
\value{
string of cleaned characters forming a sentence or phrase.
}
\description{
Converts various type of spaces into "normal" spaces.
}
\details{
One of several text cleaning functions. The function replaces various
types of spaces with ASCII spaces that are breakable. The reason for this
function is to replace non-breakable space with standard, breakable spaces.
}
\examples{
\dontrun{
  comment <- "All day I see class mates and send e-mails.   I miss East Lansing."

  convert.spaces(text=comment)
}

}
\seealso{
Other "text cleaning": 
\code{\link{clean.cntrl}()},
\code{\link{extract.emails}()}
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\concept{"text cleaning"}
