% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.termCode}
\alias{convert.termCode}
\title{Convert Four-Digit Term Codes}
\usage{
convert.termCode(term.code, term.type = "full")
}
\arguments{
\item{term.code}{Four-digit code indicating the semester and year of the term.}

\item{term.type}{Indicate if the full or short semester (or quarter) designation
is returned; default: \code{"full"}.}
}
\value{
string term and four-digit year
}
\description{
Converts four-digit term codes to semester or quarter names with
four-digit years.
}
\details{
Within CampusSolutions the term code is a four digit representation
of the century (position 1), the two-digit calendar year (positions 2 and 3),
and the term (position 4). The following is the translation between digits
and four-digit years and term types.\tabular{cc}{
   \strong{Position} \tab \strong{Information} \cr
   1 \tab Century Code \cr
    \tab (1 = 19xx & 2 = 20xx) \cr
   2 & 3 \tab Two-Digit Calendar Year \cr
   4 \tab Term Indicator (see below) \cr
}


The information in the following table is available in the constant
\link{term.translation}.\tabular{ccc}{
   \strong{Term abbreviation} \tab \strong{Full Term Name} \tab \strong{Short Term Name} \cr
   1 \tab Winter Quarter (WinterQ) \tab WQ \cr
   2 \tab Spring \tab SS \cr
   3 \tab Spring Quarter (SpringQ) \tab SQ \cr
   5 \tab Summer \tab US \cr
   6 \tab Summer Quarter \tab UQ \cr
   8 \tab Fall \tab FS \cr
   9 \tab Fall Quarter (FallQ) \tab FQ \cr
}
}
\examples{
term.code <- "2208"

convert.termCode(term.code, term.type="full")
# [1] "Fall 2020"

convert.termCode(term.code, term.type="short")
# [1] "FS 2020"

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
