% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_text.R
\name{convert.toASCII}
\alias{convert.toASCII}
\title{Convert to ASCII}
\usage{
convert.toASCII(comment)
}
\arguments{
\item{comment}{A string of words forming a sentence or phrase.}
}
\value{
string of cleaned characters forming words
}
\description{
Converts non-local characters to ASCII
}
\details{
Convert \emph{\strong{latin1}} encoded characters to ASCII
}
\examples{
comment <- "All day I see class mates and send e-mails.   I miss East Lansing."

convert.toASCII(comment=comment)

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
