% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.YN2TF}
\alias{convert.YN2TF}
\title{Convert Yes-No indicators to TRUE-FALSE}
\usage{
convert.YN2TF(YN.string)
}
\arguments{
\item{YN.string}{string with "Yes", "No" indicators such as "Y", "N",
"yes", "no", "1", or "0"}
}
\value{
logical vector
}
\description{
Often True-False data is returned as a vector of Ys and Ns or 1s
and 0s. These vectors often contain blanks or \code{NA}s that cannot, and should
not, be converted into Falses. This function converts
\itemize{
\item "Y" -->> \code{TRUE}
\item "N" -->> \code{FALSE}
\item "Yes" -->> \code{TRUE}
\item "No" -->> \code{FALSE}
\item "1" -->> \code{TRUE}
\item "0" -->> \code{FALSE}
\item 1 -->> \code{TRUE}
\item 0 -->> \code{FALSE}
}

and leave blanks and \code{NA}s in the vector.
}
\examples{
YN.string <- c("Y", "N", "1", "0", "yes", "no", NA, NA)
# [1]  TRUE FALSE  TRUE FALSE    NA    NA

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
