% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{course.type}
\alias{course.type}
\title{Determine the Course Type}
\usage{
course.type(crse_code)
}
\arguments{
\item{crse_code}{single character vector with the course code of interest.}
}
\value{
logical
}
\description{
Determine if a course is a preparation (aka prep), honours (H),
multi-diversity (I), national diversity (N), international and multicultural
diversity (D), or writing (W) course based
on the course code (\code{crse_code}). Theese designations are based on the
\href{https://reg.msu.edu/Read/UCC/Courselistings.pdf}{"Definitions of Course Characteristics"}
document provided on the Office of the Registrar's website.
}
\examples{
crse_code <- "1831"
course.type(crse_code)
# "prep"

crse_code <- "1838"
is.prep(crse_code)
# ""

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
