% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{create.search.string}
\alias{create.search.string}
\title{Construct the Search String}
\usage{
create.search.string(search.terms)
}
\arguments{
\item{search.terms}{vector of strings to use as the query (aka "pattern") string.}
}
\value{
string with the terms to search the text string of interest, \emph{e.g.},
\code{"\\\\bTERM1\\\\b|\\\\bTERM2\\\\b"}.
}
\description{
Construct the search string (aka "pattern") based on provided
search terms. The \code{\link[=get.unique.terms]{get.unique.terms()}} and \code{\link[=get.bigrams.oi]{get.bigrams.oi()}} functions use this
function. The function is a simple wrapper for the \code{\link[=paste]{paste()}} function that
surrounds each search term with \verb{\\\\b} and separates multiple terms with
pipes \code{|}.
}
\examples{
\dontrun{
search.terms <- c("dog", "cat", "fish")
create.search.string(search.terms)
"\\\\bdog\\\\b|\\\\bcat\\\\b|\\\\bfish\\\\b"
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
