% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-grades.R
\name{crossover.pseudo.grades}
\alias{crossover.pseudo.grades}
\title{Crossover Pseudo-Grade}
\usage{
crossover.pseudo.grades(grades, crossover.point = "random")
}
\arguments{
\item{crossover.point}{}
}
\value{

}
\description{
add the description of making the data
}
\examples{
\dontrun{
 crossover.pseudo.grades(grades=grades, crossover.point=5)
 crossover.pseudo.grades(grades=grades, crossover.point="random")
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
