% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_CONSTANTS.R
\docType{data}
\name{deg2rad}
\alias{deg2rad}
\alias{earth.radius.miles}
\alias{earth.radius.km}
\title{Calculate Latitude-Longitude Distance Constants}
\format{
numeric (double)

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
deg2rad

earth.radius.miles

earth.radius.km
}
\description{
Constants needed to calculate the distance between two latitude-
longitude points on a globe, aka earth.
\itemize{
\item \code{deg2rad} | Degrees to Radians: \code{0.0174532925199433} for the
conversion of degrees (specifically latitude and longitude values) to
radians. Needed for calculating the distance between two locations on the
globe using latitude-longitude values.
}

\loadmathjax
\mjdeqn{ deg2rad = \frac{\pi}{180} }{deg2rad = pi/180}
\itemize{
\item \code{earth.radius.miles} | Earth's Radius: \code{3963.17} miles.
\item \code{earth.radius.km} | Earth's Radius: \code{6373} kilometers.
}
}
\section{Functions}{
\itemize{
\item \code{earth.radius.miles}: Earth's Radius

\item \code{earth.radius.km}: Earth's Radius
}}

\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\keyword{datasets}
