% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_text.R
\name{extract.emails}
\alias{extract.emails}
\title{Extract Email Address(es) From String}
\usage{
extract.emails(emails.string, domains = "msu.edu")
}
\arguments{
\item{emails.string}{string with emails.}

\item{domains}{string with the domains; \emph{e.g.}, "msu.edu".}
}
\value{
string of emails
}
\description{
Extract all email address(es) with a specific domain (\emph{e.g.},
msu.edu, gmail.com, hotmail.com, etc.)
}
\examples{
extract.emails(emails.string="emilio.dork@gmail.com 2.that.dork@egr.msu.edu
dork@egr.msu.edu 517565656325@162.123 thatDork@msu.edu", domains="msu.edu")
# [1] "2.that.dork@egr.msu.edu" "thatdork@msu.edu"

}
\seealso{
Other "text cleaning": 
\code{\link{clean.cntrl}()},
\code{\link{convert.spaces}()}
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\concept{"text cleaning"}
