% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_datasets.R
\name{extract.unique}
\alias{extract.unique}
\title{Extract Unique Values from Each Column}
\usage{
extract.unique(dataset, cell.blank.tf, cell.NA.tf, size = 3)
}
\arguments{
\item{dataset}{\code{tibble} or \code{data.frame} of interest}

\item{cell.blank.tf}{constructed within \code{\link[=dataset.summary]{dataset.summary()}} using the
command \code{purrr::map_dfc(dataset, is.BLANK)} and returns a tibble the same
dimensions as \code{dataset} and indicates whether or not each cell is blank.
\code{TRUE} indicates the cell is blank and \code{FALSE} indicates it contains some
type of value. \code{NA}s are \code{\link[dplyr:coalesce]{dplyr::coalesce()}}d into \code{FALSE}.}

\item{cell.NA.tf}{constructed within \code{\link[=dataset.summary]{dataset.summary()}} using the
command \code{purrr::map_dfc(dataset, is.na)} and returns a tibble the same
dimensions as \code{dataset} and indicates whether or not each cell is blank.
\code{TRUE} indicates the cell contains a \code{NA} and \code{FALSE} indicates it contains some
type of value.}

\item{size}{integer value indicating the number examples to return}
}
\value{
tibble of examples
}
\description{
Given a \code{tibble} or \code{data.frame} extract unique examples for
each column. This function is not publically available, but maybe a future
version will be.
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
