% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_PCA.R
\name{extract.VarExplain}
\alias{extract.VarExplain}
\title{Extract Variance Explained (PCA)}
\usage{
extract.VarExplain(PCAmodel, nDigits.label = 1)
}
\description{
Extract Variance Explained (PCA)
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
