% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filename-times}
\alias{filename-times}
\alias{file.datetime}
\alias{file.date}
\alias{file.time}
\title{File Name Date and Time}
\usage{
file.datetime(date.time = NULL)

file.date(date.time = NULL)

file.time(date.time = NULL)
}
\arguments{
\item{date.time}{string extracted using \code{\link[base:Sys.time]{base::Sys.time()}} or passing \code{NULL}
will result in the time the \code{Sys.time()} function is called being used.

\emph{\strong{NOTE}}: Do \emph{\strong{NOT}} pass \code{NULL} to the individual date and time
functions. Instead, call \code{file.datetime()} and split the string using
\code{\link[base:strsplit]{base::strsplit()}} in the form \code{strsplit(x=file.datetime(), split="_")}
to obtain the date and time.}
}
\value{
string with the date, time, or date_time in the format noted above.
\itemize{
\item For \code{file.date()}, the date "Jan232021"
\item For \code{file.time()}, the time "1453"
\item For \code{file.datetime()}, the date and time "Jan232021_1453"
}
}
\description{
Construct date and time components for a file name. Ideally, this
function is called when the script (or main function) is initiated.
\itemize{
\item \code{file.date()}: returns the date in "\%b\%d\%Y" format; \emph{e.g.}, "Jan232021"
\item \code{file.time()}: returns the time in "\%H\%M" format; \emph{e.g.}, "1453"
\item \code{file.datetime()}: returns the date and time in the "\%b\%d\%Y_\%H\%M" format;
\emph{e.g.}, "Jan232021_1453"
}
}
\examples{
date.time <- as.POSIXlt("2021-01-23 14:53:13 EST")

file.date(date.time=date.time)
# "Jan232021"

file.time(date.time=date.time)
# "1453"

file.datetime(date.time=date.time)
# "Jan232021_1453"

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
\concept{filenames}
