% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_datasets.R
\name{find.duplicate.cols}
\alias{find.duplicate.cols}
\title{Find Duplicate Columns}
\usage{
find.duplicate.cols(data, data.md5s)
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame} of interest}

\item{data.md5s}{the calculated md5 hashes}
}
\value{
data.frame with the above information
}
\description{
Identifies and returns duplicate columns. The resulting data.frame
includes the following:#'
\itemize{
\item column.name: column name
\item duplicate: logical indicating if the column is duplicate. FYI: only duplicate
columns are returned
\item duplicate.colName: duplicate columns
}

\emph{\strong{Not a publicly available function at this time.}}
}
\examples{
data <- tibble(first.name=c("Alice", "Bob", "Carl", "Debbie"),
                  last.name=c("Masters", "Roberts", "Roberts", "Smith"),
                  pref.name=c("Alice", "Bob", "Carl", "Debbie"),
                  role=c("data", "data", "coding", "data"),
                  job=c("data", "data", "coding", "data"),
                  expert=c("data", "data", "coding", "data"),
                  building=c("Chemistry", "Biochemistry", "Chemistry", "Statistics"))
find.duplicate.cols(data=data)
#  column.name duplicate.tf    duplicate.colNames
#1  first.name         TRUE first.name, pref.name
#2   pref.name         TRUE first.name, pref.name
#3        role         TRUE     role, job, expert
#4         job         TRUE     role, job, expert
#5      expert         TRUE     role, job, expert

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
