% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_project.R
\name{framework.project}
\alias{framework.project}
\title{Create Project Framework}
\usage{
framework.project()
}
\value{
Nothing is returned to the R session. The following \emph{\strong{might}} be created
in the current working directory:
\itemize{
\item \verb{00_README.md}
\item \code{./DATA}
\item \code{./DOCUMENTS}
\item \code{./IMAGES}
\item \code{./PRESENTATIONS}
\item \code{./REFERENCES}
\item \code{./REPORTS}
\item \code{./RESULTS}
}
}
\description{
The creation of the base setup for a project. The function checks
to see if a file with \code{readme} in its name and if any directories (aka folders)
with the words data, document, image, and presentation are present within the
current working directory. The \code{README} and the folders noted below are only
created if the not already present. The script indicates what is and is not
being created.
\itemize{
\item \verb{00_README.md}: Provide general information about the project, what is
being worked on, and what is completed.
\item \code{./DATA}: Where files containing the data for the study reside.
\item \code{./DOCUMENTS}: A location for reports, articles, and presentations
related to the project and are often provided by colleagues.
\item \code{./IMAGES}: Where all the images created within the scripts are written.
\item \code{./PRESENTATIONS}: Where the presentations and other reports created for the
project from the performed analyses.
\item \code{./REFERENCES}: A location for reports, articles, and presentations
related to the project and are often provided by colleagues.
\item \code{./REPORTS}: Where the presentations and other reports created for the
project from the performed analyses.
\item \code{./RESULTS}: Where the presentations and other reports created for the
project from the performed analyses.
}

The \code{README} file can work as an extension of your notebook by containing links
to websites and code snippets (short pieces of code to perform specific tasks).
The \code{README} file is constructed as a Markdown file and allows you to provide
simple text formatting, the potential to be converted into a rendered version
(PDF, webpage aka HTML, and Word document), and is human readable and easily
edited. The leading double zeros (\code{00}) ensure the \code{README} file is always at
the top of the directory when the files are ordered by name in descending order.

This function should be run after creating the R project.
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
