% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{get.bigrams.oi}
\alias{get.bigrams.oi}
\title{Find n-grams with Keywords}
\usage{
get.bigrams.oi(data, col.oi, search.terms)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble} with an n-gram to search.}

\item{col.oi}{string with the column name of to search.}

\item{search.terms}{string with the keyword(s) of interest. Can be a single word
(\emph{e.g.}, "cat") or multiple words (\emph{e.g.}, \code{c("dog", "cat")}).}
}
\value{
\code{data.frame} or \code{tibble} (depending on what is provided) with the n-gram
phrases connected to the provided \code{search.terms} (aka keyword(s)).
}
\description{
When making network map/plot of connected terms, the need to remove
cluster of terms with three or two terms is often desired. This function
identifies all n-grams connected to the user provided keyword(s).
}
\examples{
\dontrun{
popular.DATA <- get.bigrams.oi(data=bigrams.n3,
                               col.oi="gram.2",
                               search.terms=c("development", "learning", "health", "research"))
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
