% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{get.unique.terms}
\alias{get.unique.terms}
\title{Get Unique Terms}
\usage{
get.unique.terms(string.oi, search.terms)
}
\arguments{
\item{string.oi}{vector of strings to search.}

\item{search.terms}{vector of strings to use as the query (aka "pattern") string.}
}
\value{
string with the terms related to the query
}
\description{
Using the provided search terms, identify the phrases containing
the search terms, and return all the related terms.
}
\examples{
\dontrun{
keywords.comp.sim <- compare.dataset(data)

keywords.heatmap.data <- keywords.plot.data(keywords.similarities=keywords.comp.sim)
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
