% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{has.keywords}
\alias{has.keywords}
\title{Keywords Present}
\usage{
has.keywords(data, comment.col, keywords.tb)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with the comments of interest}

\item{comment.col}{string indicating the comments of interest; \emph{e.g.}, \code{"comments.clean"}}

\item{keywords.tb}{\code{tibble} with the keywords and area}
}
\value{
\code{tibble} (or \code{data.frame}) with the original data plus a column
for each provided area. Each \emph{area} column has a \code{TRUE} or \code{FALSE} for
each row indicating if the comment had a \emph{keyword} present in the comment.
}
\description{
Using either the \link{keywords} provided within \link{theHUB} or a \code{tibble}
with a collection of user defined terms and areas. The \code{tibble} needs to
have each keyword in the \emph{keyword} column and the corresponding area in the
\emph{area} column.
}
\details{
Convert the \link{keywords} to a \code{tibble} with a row for each \emph{area} using
the following command.

\if{html}{\out{<div class="sourceCode">}}\preformatted{keywords
 A tibble: 206 x 2
  keyword    area
  <chr>      <chr>
1 adviser    acadSupport
2 advising   acadSupport
3 advisor    acadSupport
4 assignment assessment

keywords.tb <- dplyr::group_by(keywords, area) |>
  dplyr::summarise(query=paste(keyword, collapse="|"))

keywords.tb
# A tibble: 16 x 2
  area          query
  <chr>         <chr>
1 acadSupport   adviser|advising|advisor
2 assessment    assignment|...
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library(dplyr)
keywords.tb <- group_by(keywords, area) |>
    summarise(query=paste(keyword, collapse="|"))

fs20.comments <- has.keywords(data=fs20.comments,
                              comment.col="comments.clean",
                              keywords.tb=keywords.tb)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
