% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ID.convert}
\alias{ID.convert}
\title{Convert Collection of Numerical IDs}
\usage{
ID.convert(ID, deID.value = 12345)
}
\arguments{
\item{ID}{string with ID or collection of IDs}

\item{deID.value}{numeric value, preferably an integer, to add to the
numeric ID; default: \code{12345}. Values are converted from string to numeric
(real) back to string to avoid the \code{.Machine$integer.max} (\verb{2,147,483,647})
problem.}
}
\value{
string vector
}
\description{
Given a collection of numerical IDs, systematically change their
values by a user defined value. \emph{\strong{This method is inferior to the
\code{\link[=ID.random]{ID.random()}} method. No judgement.}}
}
\examples{
IDs <- c("1234500000", "12345", NA, 12345, 1234.56, "puzzle", "00011111")
ID.convert(ID=IDs, deID.value=12345)
[1] "1234512345" "24690"      NA           "24690"      "13579.56"   NA           "00023456"
Warning message:
In ID.convert(ID = IDs, deID.value = 12345) : NAs introduced by coercion

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})

Seth Walker \email{walker893@msu.edu}
(\url{https://github.com/walker893})
}
