% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.BLANK}
\alias{is.BLANK}
\title{Is String Blanks?}
\usage{
is.BLANK(string)
}
\arguments{
\item{string}{or vector of strings to evaluate.}
}
\value{
logical
}
\description{
Determine which strings within a vector of strings is either
a string with no characters or a string comprised entirely of blanks.
This function uses \code{\link[stringr:str_detect]{stringr::str_detect()}} and the regular expression
\code{"^[[:space:]]*$"} to determine if the string contains no characters or
all blanks.
}
\examples{
string <- c("", " ", "  ", "   ", " abc", "abc ", " abc ")
is.BLANK(string=string)
# [1]  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
