% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.honours}
\alias{is.honours}
\title{Honours (or Honors) Course?}
\usage{
is.honours(crse_code)
}
\arguments{
\item{crse_code}{single character vector with the course code of interest.}
}
\value{
logical
}
\description{
Determine if a course is an honours course based on the course
code (\code{crse_code}). The fourth character is an "H". This designation is
based on the \href{https://reg.msu.edu/Read/UCC/Courselistings.pdf}{"Definitions of Course Characteristics"}
document provided on the Office of the Registrar's website.
}
\examples{
crse_code <- "183H"
is.honours(crse_code)

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
