% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{is.none}
\alias{is.none}
\title{No Comment Present}
\usage{
is.none(comment)
}
\arguments{
\item{comment}{string with the comment}
}
\value{
logical indicating if the comment is a common non-response
}
\description{
Indicates if the presented comment was intended to be a "none-
comment" by the author. Steps to determine if the user supplied comment is
considered a non-comment:
\itemize{
\item convert string to lower case
\item count the number of words
\item count the number of characters
\item if the string is a single word OR the number of characters is zero (0)
\itemize{
\item remove all punctuation
\item if the single word is: "", "na", "no", "none", "nothing" indicate the
comment was intended to be a "non-comment"
}
}
}
\examples{
 is.none("today is awesome!")
 # FALSE

 is.none("NA")
 # TRUE

 is.none("N/A")
 # TRUE

 is.none("")
 # TRUE

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
