% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.prep}
\alias{is.prep}
\title{Course Prep?}
\usage{
is.prep(crse_code)
}
\arguments{
\item{crse_code}{single character vector with the course code of interest.}
}
\value{
logical
}
\description{
Determine if a course is a preparation, aka prep, course based
on the course code (\code{crse_code}). The fourth character, in this case a
digit is five (5) or less. This designation is based on the
\href{https://reg.msu.edu/Read/UCC/Courselistings.pdf}{"Definitions of Course Characteristics"}
document provided on the Office of the Registrar's website.
}
\examples{
crse_code <- "1831"
is.prep(crse_code)
# TRUE

crse_code <- "1838"
is.prep(crse_code)
# FALSE

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
