% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{keywords.plot.data}
\alias{keywords.plot.data}
\title{Construct Keyword Similarity Plots}
\usage{
keywords.plot.data(keywords.similarities)
}
\arguments{
\item{keywords.similarities}{results of \code{\link[=compare.dataset]{compare.dataset()}}.}
}
\value{
\code{tibble} with the data to construct a heatmap
}
\description{
Construct the pairwise heatmap data for keyword area overlap
}
\examples{
\dontrun{
keywords.comp.sim <- compare.dataset(data)

keywords.heatmap.data <- keywords.plot.data(keywords.similarities=keywords.comp.sim)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
