% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_CONSTANTS.R
\docType{data}
\name{keywords}
\alias{keywords}
\alias{keyword.areas}
\title{Keywords and Areas of Interest}
\format{
A tibble with 206 rows and 2 columns:
\itemize{
\item keyword: Word
\item area:
}
}
\source{
List of keywords and associated areas as determined by Emilio. Please
feel free to offer suggestions to expand keywords and areas covered.
}
\usage{
keywords
}
\description{
A collection of keywords (206 unique keywords) and associated
areas (16)
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
\keyword{datasets}
