% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{LatLonDistances}
\alias{LatLonDistances}
\title{Latitude-Longitude Distances}
\usage{
LatLonDistances(lat1, long1, lat2, long2, distance.units = "miles")
}
\arguments{
\item{distance.units}{string indicating if the resulting distance should be
miles or kilometers; default is \code{"miles"}.}

\item{point1}{vector with two numeric values for the first pair of
latitude-longitude coordinates;
\emph{e.g.}, c(latitude, longitude), c(42.331177, -83.046023)  ## Detroit, Michigan}

\item{point2}{vector with two numeric values for the first pair of
latitude-longitude coordinates;
\emph{e.g.}, c(latitude, longitude), c(42.737652, -84.483788)  ## East Lansing, Michigan}
}
\value{
distance between the two points taking into consideration the curvature of the earth
}
\description{
Calculate distance between points on the globe
}
\details{
long description with what the function does and returns
}
\examples{
point1 <- c(42.331177, -83.046023)  ## Detroit, Michigan
point2 <- c(42.737652, -84.483788)  ## East Lansing, Michigan

LatLonDistances(point1, point2, distance.units="miles")
# 78.48957

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
