% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert.barplot}
\alias{likert.barplot}
\title{Likert (and Likert-like) Barplot}
\usage{
likert.barplot(data, QoI, value.range)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with Likert data converted to integers.
Use the \code{\link[=convert.fromto]{convert.fromto()}} function to convert text responses to integers.
\emph{\strong{Note}}: Ensure the integers are positive and non-zero.}

\item{QoI}{string with the Question of Interest (the rows)}

\item{value.range}{range of possible Likert values}
}
\value{
ggplot2 graphics object
}
\description{
Constructs a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} barplot for Likert and Likert-like results
with each option displayed as an individual, horizontal MSU dark-green bars.
The percentage of responses are noted on the bar as white text.

This function is called by \code{\link[=likert.plot.matrix]{likert.plot.matrix()}} but can be called individually.
}
\examples{
\dontrun{
likert.barplot(data=fs20.likert, QoI="Q1")
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
