% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert.barplot.stacked}
\alias{likert.barplot.stacked}
\title{Likert (and Likert-like) Stacked Barplot}
\usage{
likert.barplot.stacked(
  data,
  colour.palette = msu.palette,
  legend.position = "bottom",
  label.col = "Label.long",
  label.size = 4,
  display.minimum = 3
)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with Likert data converted to integers.
Use the \code{\link[=convert.fromto]{convert.fromto()}} function to convert text responses to integers.
\emph{\strong{Note}}: Ensure the integers are positive and non-zero.}

\item{colour.palette}{the colour palette of interest; default: \code{"msu.palette"}}

\item{legend.position}{to legend's position; default: \code{"bottom"}}

\item{label.col}{string with the column containing the labels. Setting to \code{NULL}
will result in no labels.}

\item{label.size}{label sizes; default: \code{4}}

\item{display.minimum}{numerical value indicating the small number of responses
to display a label; default: \code{3}. Thus, responses with 2 or less responses
are displayed within the stacked barplot but are not labeled.}
}
\value{
ggplot2 graphics object
}
\description{
Constructs a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} barplot for Likert and Likert-like results
with each option displayed as an individual, horizontal MSU dark-green bars.
The percentage of responses are noted on the bar as white text.

This function is called by \code{\link[=likert.plot.matrix]{likert.plot.matrix()}} but can be called individually.
}
\examples{
\dontrun{
likert2int.tb <- tibble::tibble(phrase=c("hated it!", "meh", "loved it!"),
                                integer=c(-1, 0, 1))
make.likert.barplot.data(data, likert2int.tb)
likert.barplot.stacked(data,
                       colour.palette=msu.palette,
                       legend.position="bottom",
                       display.minimum=3)
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
