% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert.heatmap}
\alias{likert.heatmap}
\title{Construct Likert (and Likert-like) Pairwise Heatmap}
\usage{
likert.heatmap(data, QoI, Qcompared, value.range)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with Likert data converted to integers.
Use the \code{\link[=convert.fromto]{convert.fromto()}} function to convert text responses to integers.
\emph{\strong{Note}}: Ensure the integers are positive and non-zero.}

\item{QoI}{string with the Question of Interest (the rows)}

\item{Qcompared}{string with the Question being Compared to (the columns)}

\item{value.range}{range of possible Likert values}
}
\value{
ggplot2 graphics object
}
\description{
Constructs a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} heatmap for Likert and Likert-like results
with each option for the question of interest (\code{QoI}) represented on the row
while the columns are the question being compared (\code{Qcompared}). Each tile
(square) is shaded based on the percent overlap and the percentage of
overlapping responses are noted as MSU dark-green text on a white box.

This function is called by \code{\link[=likert.plot.matrix]{likert.plot.matrix()}} but can be called individually.
}
\examples{
\dontrun{
likert.heatmap(data=fs20.likert, QoI="Q1", Qcompared="Q2")
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
