% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert.plot.matrix}
\alias{likert.plot.matrix}
\title{Likert (and Likert-Like) Pairwise Analysis Plot}
\usage{
likert.plot.matrix(data, value.range, title, questions, footnotes = NULL)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) of Likert and categorical data with
responses in integer form.}

\item{value.range}{XXXXXXX}

\item{title}{string with the title for the plot; \emph{e.g.}, "Likert Pairwise Analysis"}

\item{questions}{vector of strings with shortened versions of the questions
used to label the plots; \emph{e.g.}, c("Q1. Overall", "Q2. MSU Leadership Communication").
Use \verb{\\n} to denoted line breaks (aka carriage returns).}

\item{footnotes}{vector of strings to represent what the numeric (integer)
values represent. Use \verb{\\n} to denoted line breaks (aka carriage returns).}
}
\value{
\link[ggplot2:ggplot2-package]{ggplot2::ggplot2} object constructed via \link[cowplot:cowplot-package]{cowplot::cowplot}
}
\description{
Graphical representation of pairwise interactions between Likert
and Likert-like responses. Likert-like responses are survey questions where
the respondent can only select one response.
}
\examples{
\dontrun{
title <- "Likert & Likert-like Comparison"
questions <- c("Q1. Do you like everything bagels?", "Q2. Do you drink coffee?")
footnotes <- c("Q1. 1='Not a lot', 2='Meh', 3='Very much';\nQ2. 1=Never, 2=Some times, 3='Always'")

likert.plot.matrix(data=select(fs20.data, Q1, Q2),
                   title=title,
                   questions=questions,
                   footnotes=footnote)
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
