% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{lonlat2region}
\alias{lonlat2region}
\title{Longitude-Latitude to Region}
\usage{
lonlat2region(longitude, latitude, gadm.mapfile)
}
\arguments{
\item{longitude}{numeric value indicating the longitude}

\item{latitude}{numeric value indicating the latitude}

\item{gadm.mapfile}{string with the path and file name of the GADM world
mapfile. Any mapfile can be used, but for datasets containing geospatial
location data from anywhere on the planet, it is advantageous to use the
entire world map. Obtain the mapfile of interest from
\url{https://gadm.org}. The world GADM data is available from
\link{https://biogeo.ucdavis.edu/data/gadm3.6/gadm36_gpkg.zip}
(https://biogeo.ucdavis.edu/data/gadm3.6/gadm36_gpkg.zip).
Ensure the \href{https://gadm.org/formats.html}{GeoPackage format} is downloaded.
The compressed complete world database is approximately 1.6 GBs and the
unzipped file is 3.81 GBs. For example, \verb{~/mapfiles/gadm36_levels.gpkg}

Individual country maps are available from
\url{https://gadm.org/maps.html}.}
}
\value{
tibble with the country.code, country.name, region, and subregion.
}
\description{
Identify the region (and subregion) of a geospatial location
position using longitude and latitude.
}
\details{
Based on the provided longitude and latitude, the corresponding
country, region, and subregion are identified. Positions outside of a
country return \code{NA}s for the country.code, country.name, region, and
subregion. Instances where \code{NA} is provided for longitude and latitude,
also returns \code{NA} values.

When using the world GADM database, loading the rather large world database
takes time and, thus, the function appears to be slow.

This function is based on the \href{https://stackoverflow.com/questions/8751497/latitude-longitude-coordinates-to-state-code-in-r}{"Latitude Longitude Coordinates to State Code in R"}
stackoverflow response by
\href{https://stackoverflow.com/users/980833/josh-obrien}{Josh O'Brien}.
}
\examples{
\dontrun{
mapfile.world <- "~/mapfiles/gadm36_levels.gpkg"
lonlat2region(longitude, latitude, gadm.mapfile=mapfile.world)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
