% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-grades.R
\name{make.assignment.params}
\alias{make.assignment.params}
\title{Construct Assignment Parameters}
\usage{
make.assignment.params(n.students, n.assignments, max.points)
}
\arguments{
\item{n.students}{number of students in the course; default: 15}

\item{n.assignments}{number of assignments or assessments; default: 10}

\item{max.points}{maximum number of points for all assignments; default: 25}
}
\value{
\code{tibble} with the input parameters for make.pseudo.grades
}
\description{
add the description of making the data
}
\examples{
assignment.params <- make.assignment.params(n.students=5,
                                            n.assignments=5,
                                            max.points=25)
# # A tibble: 5 × 3
# n  size  prob
# <dbl> <dbl> <dbl>
# 1     5    25  0
# 2     5    25  0.25
# 3     5    25  0.5
# 4     5    25  0.75
# 5     5    25  1

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
