% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make.bins}
\alias{make.bins}
\title{Bin Values}
\usage{
make.bins(x, bins, bin.names)
}
\arguments{
\item{x}{vector of (likely) integer values representing age}

\item{bins}{vector of bins for use with \code{\link[base:findInterval]{base::findInterval()}}}

\item{bin.names}{vector of bin names. \emph{\strong{NEEDS}} to be the same
length as \code{bins}}
}
\value{
string with interval values
}
\description{
Bin and name a collection of values.
}
\details{
Commonly used to bin ages and survey responses that are continuous;
\emph{e.g.}, \verb{1, 3, 1, 2, 5, 4}. Values that are not bin-able return a \verb{"Not Provided}"
classification.
}
\examples{
set.seed(13)
age <- c(NA, sample(1:10, 20, replace=TRUE), NA)
age.bins <- c(0, 2, 4, 6, 8, 10)
age.bin.names <- c("0", "2-3", "4-5", "6-7", "8-9", "10")
make.bins(x=age, bins=age.bins, bin.names=age.bin.names)

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
