% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make.crsesect}
\alias{make.crsesect}
\title{Make Course-Section Designation}
\usage{
make.crsesect(course, sctn_code)
}
\arguments{
\item{course}{string containing the course designation; a combination of
\code{subj_code} and \code{crse_code}. For example, \code{"PKG 491"}.}

\item{sctn_code}{string (or integer) with the section code.}
}
\value{
string with the subject code, course code, and section code to form
a course-section designation in the same format as the \code{crsesect_firstday},
\code{crsesect_qrtrterm}, and \code{crsesect_endterm} columns/fields.
}
\description{
Constructs the course-section designation to match the course-section
description in the \code{crsesect_firstday}, \code{crsesect_qrtrterm}, and \code{crsesect_endterm}
columns/fields
}
\examples{
course <- "PKG 491"
sctn_code <- "6"

make.crsesect(course, sctn_code)
# "PKG 491-006"

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
