% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_donut.R
\name{make.donut.data}
\alias{make.donut.data}
\title{Construct Donut Plot Data}
\usage{
make.donut.data(
  data,
  col.oi,
  groupby = NULL,
  category.order = "count",
  levels.rev = FALSE,
  r.inner = 4,
  r.outer = 6
)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with the column of interest. \emph{\strong{NOTE}}:
Do \strong{NOT} use counted data.}

\item{col.oi}{string with column of interest. Only provide \strong{ONE} column name.}

\item{levels.rev}{logical indicating if the order of the categories should be
reversed.}

\item{r.inner}{numeric value defining the inner radius of the donut; default: 4}

\item{r.outer}{numeric value defining the outer radius of the donut; default: 6}
}
\value{
\code{tibble} with the
}
\description{
add the description of making the data
}
\examples{
\dontrun{
  donut.DATA <- make.donut.data(data, col.oi, groupby=NULL,
                                category.order="count", levels.rev=FALSE,
                                r.inner=4, r.outer=6)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
