% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{make.fps}
\alias{make.fps}
\title{Make Fingerprints}
\usage{
make.fps(data, id.col, cols.oi)
}
\arguments{
\item{data}{data.frame or tibble containing raw data to convert into fingerprints}

\item{id.col}{string indicating the identifying column}

\item{cols.oi}{vector of integers or strings indicting the columns of interest}
}
\value{
tibble of fingerprints with integer (count) or float values
}
\description{
Construct fingerprints for a collection of samples from the provided
integer and float values.
}
\examples{
\dontrun{
  library(tidyverse)
  where2eat <- tibble::tibble(what=c("BBQ", "Burgers", "Pizza"),
                              loc=c(1,2,3),
                              size=c(4,2,6),
                              cost=c(10,9,8))
  make.fps(data=where2eat, id.col="what", cols.oi=c("loc", "size", "cost"))
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
