% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{make.likert.barplot.data}
\alias{make.likert.barplot.data}
\title{Construct Likert Stacked, Barplot Data}
\usage{
make.likert.barplot.data(data, likert2int.tb)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with Likert data converted to integers.
Use the \code{\link[=convert.fromto]{convert.fromto()}} function to convert text responses to integers.
\emph{\strong{Note}}: Ensure the integers are positive and non-zero.}

\item{likert2int.tb}{\code{tibble} (or \code{data.frame}) with the conversion between the
\emph{\strong{long}} description and the \emph{\strong{integer}} value. See \code{\link[=convert.fromto]{convert.fromto()}}
for an example of the \code{tibble}'s construction.}
}
\value{
\code{tibble} (or \code{data.frame}) containing the needed count, percentage, and
xxxx to construct a Likert stacked, bar chart with \code{\link[=likert.barplot.stacked]{likert.barplot.stacked()}}.
}
\description{
Construct the Likert data for \code{\link[=likert.barplot.stacked]{likert.barplot.stacked()}}.
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
