% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{make.likert.heatmap.data}
\alias{make.likert.heatmap.data}
\title{Construct Pairwise Heatmap Data}
\usage{
make.likert.heatmap.data(data, QoI, Qcompared, value.range)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with Likert data converted to integers.
Use the \code{\link[=convert.fromto]{convert.fromto()}} function to convert text responses to integers.
\emph{\strong{Note}}: Ensure the integers are positive and non-zero.}

\item{QoI}{string with the Question of Interest (the rows)}

\item{Qcompared}{string with the Question being Compared to (the columns)}

\item{value.range}{range of possible Likert values}
}
\value{
\code{tibble} (or \code{data.frame}) containing the needed count, percentage, and
colour hex values to construct a heatmap with percentage values overlaid on
each tile of the heatmap.
}
\description{
Construct the pairwise heatmap data for \code{\link[=likert.heatmap]{likert.heatmap()}}.

This function is called by \code{\link[=likert.heatmap]{likert.heatmap()}} but can be called individually.
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
