% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_PCA.R
\name{make.PCAplot.DATA}
\alias{make.PCAplot.DATA}
\title{Make PCA Plot Data (\code{prcomp})}
\usage{
make.PCAplot.DATA(PCAmodel, PCs.oi, scale = 1, expand = 1)
}
\description{
Make PCA Plot Data (\code{prcomp})
}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
