% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-grades.R
\name{make.pseudo.grades}
\alias{make.pseudo.grades}
\title{Construct Pseudo-Grades}
\usage{
make.pseudo.grades(assignment.params)
}
\arguments{
\item{assignment.params}{tibble of grade distributions constructed by
\code{\link[=make.assignment.params]{make.assignment.params()}}.}
}
\value{
\code{tibble} with the pseudo-grades
}
\description{
Constructs pseudo-grades based on the assignment parameters
constructed by \code{\link[=make.assignment.params]{make.assignment.params()}}. Each row is a \emph{unique student}
with scores for requested number of assignments. The pseudo-grades are
assigned using the \code{\link[stats:Binomial]{stats::rbinom()}} function.
}
\examples{
set.seed(13)
make.pseudo.grades(assignment.params)
# # A tibble: 5 × 6
#   student   assign.1 assign.2 assign.3 assign.4 assign.5
#  <chr>        <int>    <int>    <int>    <int>    <int>
# 1 student.1        0        0        0        0        0
# 2 student.2        7        5        6        3       10
# 3 student.3        7       13       14       15        8
# 4 student.4       18       16       16       18       18
# 5 student.5       25       25       25       25       25

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
