% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_COLOUR-PALETTES.R
\docType{data}
\name{msu.darkGreen}
\alias{msu.darkGreen}
\alias{msu.darkGreenTints}
\alias{msu.palette}
\alias{msu.heatmap.20}
\alias{msu.heatmap.100}
\title{MSU Colour Palette}
\format{
A vector with the colours of interest.

An object of class \code{character} of length 10.

An object of class \code{character} of length 13.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101 rows and 2 columns.
}
\source{
MSU's Brand website:
\url{https://brand.msu.edu/design-visual/index.html#color}
\url{https://cal.msu.edu/about/the-marketing-communications-office/colors/}
}
\usage{
msu.darkGreen

msu.darkGreenTints

msu.palette

msu.heatmap.20

msu.heatmap.100
}
\description{
The MSU colour palette to easily construct plots. There are three
colour palettes based on the colour information found at
\url{https://brand.msu.edu/design-visual/index.html#color}.
All colours within \link{theHUB} are defined using hex-codes (also known as "web")
values.
\itemize{
\item msu.darkGreen (Primary Colour) Hex: \href{https://www.color-hex.com/color/18453b}{"#18453b"}
\item msu.darkGreenTints: Ten tints of msu.darkGreen including "#18453b" from
\href{https://www.color-hex.com/color/18453b}{https://color-hex.com/color/18453b}
\item msu.heatmap.20: Twenty tints of MSU Dark Green to the lightest tint in msu.darkGreenTints
\item msu.heatmap.100: One-hundred tints of MSU Dark Green to the lightest tint in msu.darkGreenTints
\url{https://www.color-hex.com/color/18453b}
\item msu.palette: The thirteen colours of MSU's colour palette
\itemize{
\item MSU green \href{https://www.color-hex.com/color/18453b}{"#18453b"}
\item kelly green \href{https://www.color-hex.com/color/0db14b}{"#0db14b"}
\item grey \href{https://www.color-hex.com/color/97a2a2}{"#97a2a2"}
\item orange \href{https://www.color-hex.com/color/f08521}{"#f08521"}
\item teal \href{https://www.color-hex.com/color/008183}{"#008183"}
\item blue-grey \href{https://www.color-hex.com/color/909ab7}{"#909ab7"}
\item dark grey \href{https://www.color-hex.com/color/535054}{"#535054"}
\item yellow-green \href{https://www.color-hex.com/color/d1de3f}{"#d1de3f"}
\item cream \href{https://www.color-hex.com/color/e8d9b5}{"#e8d9b5"}
\item texas-brown \href{https://www.color-hex.com/color/c89a58}{"#c89a58"}
\item split pea soup green \href{https://www.color-hex.com/color/94ae4a}{"#94ae4a"}
\item eggplant \href{https://www.color-hex.com/color/6e005f}{"#6e005f"}
\item sienna \href{https://www.color-hex.com/color/cb5a28}{"#cb5a28"}
}
}
}
\section{Functions}{
\itemize{
\item \code{msu.darkGreenTints}: Ten tints of msu.darkGreen including "#18453b"

\item \code{msu.palette}: The thirteen colours of MSU's color palette

\item \code{msu.heatmap.20}: Twenty tints of MSU Dark Green to the lightest tint in msu.darkGreenTints

\item \code{msu.heatmap.100}: One-hundred tints of MSU Dark Green to the lightest tint in msu.darkGreenTints

}}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\keyword{datasets}
