% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_PCA.R
\name{negMin.posMax.range}
\alias{negMin.posMax.range}
\title{Set Minimum Value Negative - Set Maximum Value Positive}
\usage{
negMin.posMax.range(x)
}
\description{
Set Minimum Value Negative - Set Maximum Value Positive
}
