% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_PCA.R
\name{negMin_posMax_range}
\alias{negMin_posMax_range}
\title{Set Minimum Value Negative - Set Maximum Value Positive}
\usage{
negMin_posMax_range(x)
}
\description{
Set Minimum Value Negative - Set Maximum Value Positive
}
