% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{origin2now}
\alias{origin2now}
\title{Time Since UNIX Epoch}
\usage{
origin2now(units = "mins")
}
\arguments{
\item{units}{A character string specifying the time units for the result.
Valid options are "auto", "secs", "mins", "hours", "days", "weeks".
Default is "mins". Invalid units will be reset to "mins" with a warning.}
}
\value{
A difftime object representing the time elapsed since the UNIX epoch
in the specified units.
}
\description{
Calculate time difference from UNIX epoch to current time

This function calculates the time difference between the UNIX epoch
(1970-01-01 00:00:00 UTC) and the current system time.
}
\examples{
# Get time since epoch in minutes (default)
origin2now()

# Get time since epoch in hours
origin2now("hours")

# Get time since epoch in days
origin2now("days")

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
