% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_datasets.R
\name{replace.colNames}
\alias{replace.colNames}
\title{Replace Column Names}
\usage{
replace.colNames(
  data,
  conversion.tb,
  original,
  new,
  new.cols.only = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) to have the column names replaced.}

\item{conversion.tb}{\code{tibble} (or \code{data.frame}) with the matched convertee and
converted pairs; see the example below.}

\item{original}{column in the \code{conversion.tb} containing the original column names to be converted}

\item{new}{column in the \code{conversion.tb} containing the new column names being converted to}

\item{new.cols.only}{logical indicating if \emph{\strong{ONLY}} changed columns should be returned; default: \code{FALSE}}

\item{verbose}{logical indicating if all the column names \emph{\strong{NOT}} replaced will be returned; default: \code{TRUE}}
}
\value{
a \code{tibble} (or \code{data.frame}) with column names replaced (based on user
provided parameters).
}
\description{
Replace original column names with new column names.
}
\details{
The function was initially designed to convert Likert responses to
integers, but it was quickly realized that it could easily be used for a
multitude of sins against data. The function relies on a user provided
\code{tibble} or \code{data.frame} with two columns; one with the characters to be
converted and the characters to be converted to.

The function is designed to work with \code{\link[dplyr:mutate]{dplyr::mutate()}} allowing multiple
conversions in a single command.

\emph{\strong{Note}}: Ideally, the integers are positive and non-zero.
}
\examples{
set.seed(13)
phrase2int.tb <- tibble::tibble(phrase=c("hated it!", "meh", "loved it!"),
                                integer=c(-1, 0, 1))
responses.words <- sample(x=c("hated it!", "meh", "loved it!"), size=5, replace=TRUE)
responses.integers <- sample(x=c(-1, 0, 1), size=5, replace=TRUE)

convert.fromto(responses=responses.words,
               fromto.tb=phrase2int.tb,
               from="phrase", to="integer")
# [1]  1 -1  0 -1  0

convert.fromto(responses=responses.integers,
               fromto.tb=phrase2int.tb,
               from="integer", to="phrase")
# [1] "meh"       "hated it!" "loved it!" "hated it!" "meh"

\dontrun{
replace.colNames(tibble.oi, Q1.ints=convert.fromto(responses=Q1,
                                         fromto.tb=phrase2int.tb,
                                         from="phrase", to="integer"))
}

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
