% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_CONSTANTS.R
\docType{data}
\name{SubjectCourse.codes.FS20}
\alias{SubjectCourse.codes.FS20}
\alias{SubjectCourse.codes}
\title{Subject-Course Codes for Each College Since Fall 2020}
\format{
A tibble with 986 rows and 16 columns:
\itemize{
\item coll_code: Two-digit code for the college hosting the course; \emph{e.g.}, "02".
Leading zeros (\code{0}) are requried for single-digit college codes.
\item short_name: Abbreviated or short-hand for the hosting college; \emph{e.g.},
"AG NAT RES". Presented in all capital letters.
\item full_name: Official name of the hosting college; \emph{e.g.},
"Agriculture and Natural Resources"
\item line_adr1: Street address of course, often a building on campus; \emph{e.g.},
"Morrill Hall of Agriculture"
\item city_name: City of course location, commonly "East Lansing"; \emph{e.g.}, "East Lansing"
\item state_code: State abbreviation of course location, commonly "MI" ; \emph{e.g.}, "MI"
\item zip_code: Zip code of course location, commonly "48823"; \emph{e.g.}, "48823"
\item cntry_code: Country abbreviation of course location, commonly "US" ; \emph{e.g.}, "US"
\item start_term_code: Unknown origin and purpose
\item end_term_code: Unknown origin and purpose
\item start_term_seq_id: Unknown origin and purpose
\item end_term_seq_id: Unknown origin and purpose
\item subj_code: Subject code; \emph{e.g.}, "ACC"
\item crse_code: Course code; \emph{e.g.}, "250"
\item index: Indexed value assigned to each subject-course based on the alphabetic
order of the \code{subject_code} and the order of the \code{crse_code}.
\item course: The subject code and the course number; \emph{e.g.}, "ACC 250"
}
}
\source{
\href{https://opb.msu.edu/functions/institution/index.html}{MSU Institutional Research}
}
\usage{
SubjectCourse.codes.FS20
}
\description{
A dataset of all offered courses at Michigan State University during the
fall 2020 (FS20) semester. The available courses will be updated every semester.
}
\keyword{datasets}
