% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{Tanimoto.Dist}
\alias{Tanimoto.Dist}
\title{Tanimoto Distance}
\usage{
Tanimoto.Dist(set1, set2)
}
\arguments{
\item{set1}{vector of \code{TRUE}/\code{FALSE} values OR vector of integers. See \strong{Details above}.}

\item{set2}{vector of \code{TRUE}/\code{FALSE} values OR vector of integers. See \strong{Details above}.}
}
\value{
numerical value indicating the Tanimoto distance
}
\description{
Calculate the Tanimoto distance (also known as the Jaccard
index). A value of \code{1} indicates \strong{no similarity} and a value of \code{0} indicates
\strong{perfect similarity}.
}
\details{
Either two \code{TRUE}/\code{FALSE} vectors or two \code{integer} vectors can be
provided. The \code{TRUE}/\code{FALSE} vectors \emph{\strong{must be the same lengths}} while
the integer vectors only have to indicate indices of interest in each
vector; see examples.

The Tanimoto coefficient is an indication of how similar two samples are to
each other based on number of bits overlapping (in common) in relationship
to the number of unique bits present in both samples. In this case, the
Tanimoto Distance is a measure of dissimilarity.
}
\examples{
  set1 <- c(TRUE,TRUE,FALSE,FALSE,FALSE)
  set2 <- c(TRUE,FALSE,TRUE,TRUE,TRUE)
  Tanimoto.Dist(set1, set2)
  # 0.8

  set1 <- which(set1)
  set1
  # 1 2
  set2 <- which(set2)
  set2
  # 1 3 4 5
  Tanimoto.Dist(set1, set2)
  # 0.8

}
\author{
Emilio Xavier Esposito \email{emilio.esposito@gmail.com}
(\url{https://github.com/emilioxavier})
}
