% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_CONSTANTS.R
\docType{data}
\name{term.translation}
\alias{term.translation}
\title{Term-Code Translation Table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
term.translation
}
\description{
Term-code translation table used to convert four-digit term-code
into term and four-digit year. The \code{term.translation} constant contains
the following information.\tabular{ccc}{
   \strong{Term abbreviation} \tab \strong{Full Term Name} \tab \strong{Short Term Name} \cr
   1 \tab Winter Quarter (WinterQ) \tab WQ \cr
   2 \tab Spring \tab SS \cr
   3 \tab Spring Quarter (SpringQ) \tab SpQ \cr
   5 \tab Summer \tab SU \cr
   6 \tab Summer Quarter \tab SuQ \cr
   8 \tab Fall \tab FS \cr
   9 \tab Fall Quarter (FallQ) \tab FQ \cr
}
}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\keyword{datasets}
