% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theHUB.R
\docType{package}
\name{theHUB-package}
\alias{theHUB}
\alias{theHUB-package}
\title{Collection of Learning Analytics tools to evaluate surveys, student outcomes, and student success.}
\description{
The Hub Utility Box (theHUB; a recursive backronym) is a collection of
analysis tools composed of
functions, scripts, electronic notebooks, and worked examples. Sharing these
tools with the MSU community improves the analyses of student outcomes,
student success, and survey responses. A standard set of processes and
methods along with a defined lexicon enables reproducibility, reduces the
time needed to complete analysis-based projects, and allows the quick
addition of new methods while encouraging data exploration. Our desire is
for theHUB’s functionality to grow over time, and others at MSU will
contribute new features and methods for the community. Developed and
implemented in R -- an open-source application with a robust community
(on campus and around the world) -- theHUB removes the need for a software
license while providing access to a myriad of statistical analysis and
graphing options. The documentation allows anyone to use and understand the
tools. theHUB will contain the Hub's LA v2.0 projects; currently the survey analysis
package, \emph{Survey Connections}.

\emph{Survey Connections}: Traditionally, survey analysis returns a collection of
response counts and comment summaries without considering important and
insightful relationships between questions and responses. The Survey
Connections package provides workflows for cleaning survey data and
performing Likert and Likert-like, multi-response, and comment analyses. The
package’s goal is to easily allow the discovery of relationships between
responses and the creation of cohorts for further exploration and analysis.
\emph{Analyze survey responses to gain a better understanding of the
relationships between responses and respondents}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilioxavier/theHUB/}
  \item \url{https://emilioxavier.github.io/theHUB/}
  \item \url{https://theHopefulBox.com}
  \item Report bugs at \url{https://github.com/emilioxavier/theHUB/issues/}
}

}
\author{
\strong{Maintainer}: Emilio Xavier Esposito \email{emilio.esposito@gmail.com} (\href{https://orcid.org/0000-0002-6193-0485}{ORCID}) (GitHub: https://github.com/emilioxavier)

}
\keyword{internal}
