% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{unified.terms}
\alias{unified.terms}
\title{Construct Unified Terms}
\usage{
unified.terms(comment, clean = TRUE)
}
\arguments{
\item{comment}{string of characters (usually text) to be evaluated and unified.
Alternatively, the column containing the comments can be passed to the
function as often is done via \code{\link[dplyr:mutate]{dplyr::mutate()}}}

\item{clean}{logical indicating if the function should also clean the comments
to be unified; default: \code{TRUE}. See \code{\link[=clean.comment]{clean.comment()}} for the details
on comment cleaning.}
}
\value{
vector of comments where predetermined terms are unified (and possibly
cleaned unless the user has changed the default). \emph{\strong{NOTE}}: Unified terms
are camel case. For example, "wi-fi" is returned as "WiFi" and "burn out"
is returned as "burnOut".
}
\description{
In text analysis, the concept of compounds words is difficult to
accurately capture, especially when performing term frequency analysis. To
accommodate the identification and counting of a multi-term idea/concept,
terms comprised of two or more words are combined into a unified term.
}
\examples{
comment <- "The Wi-Fi is down. My mental health is bad because  I lost my financial Aid."

unified.terms(comment=comment)
# "the WiFi is down. my mentalHealth is bad because i lost my finAid."

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
