% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_ttest.R
\name{values.t.test}
\alias{values.t.test}
\title{t-test Values}
\usage{
values.t.test(
  x,
  y,
  alternative = "two.sided",
  paired = FALSE,
  conf.level = 0.99,
  na.rm = TRUE
)
}
\arguments{
\item{alternative}{The type of t-test analysis; default: \code{"two.sided"}}

\item{paired}{\code{logical} indicating if the data is the result of a paired
sample; default: \code{FALSE}}

\item{conf.level}{The confidence level; default: \code{0.99}}

\item{na.rm}{\code{logical} indicating if \code{NA} values are removed; default: \code{TRUE}}

\item{data}{The \code{list} with the values to be compared. Each list should have
a name otherwise names will be autogenerated.}
}
\value{
t.value The t-test value

df The degrees of freedom

p.value The p-value
}
\description{
Calculate t-test and Return Key Values
}
\details{
Calculate t-test and return the key values.
}
\examples{
  values <- list(a=1:3, b=1:5, c=1:10, d=1:20)
  t.test.combinations(values)

  values <- list(a=1:3, b=1:5, c=1:10, d=1:20)
  t.test.combinations(values)


}
\author{
Emilio Xavier Esposito \email{emilio@exeResearch.com}
(\url{https://github.com/emilioxavier})
}
