% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_COLOUR-PALETTES.R
\docType{data}
\name{wisc.badgerRed}
\alias{wisc.badgerRed}
\alias{wisc.badgerRedTints}
\alias{wisc.secondary}
\alias{wisc.accent}
\alias{wisc.heatmap.20}
\alias{wisc.heatmap.100}
\title{University of Wisconsin Colour Palette}
\format{
A vector with the colours of interest.

An object of class \code{character} of length 10.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101 rows and 2 columns.
}
\source{
University of Wisconsin's Brand website:
\url{https://brand.wisc.edu/print/colors/}
}
\usage{
wisc.badgerRed

wisc.badgerRedTints

wisc.secondary

wisc.accent

wisc.heatmap.20

wisc.heatmap.100
}
\description{
The University of Wisconsin colour palette to easily construct plots.
There are three colour palettes based on the colour information found at
\url{https://brand.wisc.edu/print/colors/}. The University of
Wisconsin colour palette is available in CMYK and was converted to Hex using
\url{https://colordesigner.io/convert/cmyktohex}.

All colours within \link{theHUB} are defined using hex-codes (also known as "web")
values.
\itemize{
\item wisc.badgerRed (Badger/Cardinal Red; Primary Colour) Hex: \href{https://www.color-hex.com/color/da004c}{"#da004c"}
\item wisc.white (corresponding colour to wisc.badgerRed) Hex: \href{https://www.color-hex.com/color/ffffff}{"#ffffff"}
\item wisc.badgerRedTints: Ten tints of wisc.badgerRed including "#18453b" from
\href{https://www.color-hex.com/color/da004c}{https://color-hex.com/color/da004c}
\item wisc.heatmap.20: Twenty tints of University of Wisconsin Badger Red to the lightest tint in wisc.badgerRed
\item wisc.heatmap.100: One-hundred tints of University of Wisconsin Badger Red to the lightest tint in wisc.badgerRed
\url{https://www.color-hex.com/color/da004c}
\item wisc.secondary: The six colours of University of Wisconsin's secondary colour palette
\itemize{
\item dark red \href{https://www.color-hex.com/color/a1002f}{"#a1002f"}
\item deep red (magenta) \href{https://www.color-hex.com/color/8b0037}{"#8b0037"}
\item grey10 \href{https://www.color-hex.com/color/e6e6e6}{"#e6e6e6"}
\item grey30 \href{https://www.color-hex.com/color/b3b3b3}{"#b3b3b3"}
\item grey55 \href{https://www.color-hex.com/color/737373}{"#737373"}
\item black \href{https://www.color-hex.com/color/000000}{"#000000"}
}
\item wisc.accent: The six colours of University of Wisconsin's accent colour palette
\itemize{
\item orange \href{https://www.color-hex.com/color/ff8000}{"#ff8000"}
\item orange-yellow \href{https://www.color-hex.com/color/ffbf00}{"#ffbf00"}
\item tan \href{https://www.color-hex.com/color/e8dfa7}{"#e8dfa7"}
\item avocado green \href{https://www.color-hex.com/color/97b85f}{"#97b85f"}
\item blue-grey \href{https://www.color-hex.com/color/6b9999}{"#6b9999"}
\item dark teal \href{https://www.color-hex.com/color/386666}{"#386666"}
}
}
}
\section{Functions}{
\itemize{
\item \code{wisc.badgerRedTints}: Ten tints of wisc.badgerRed including "#da004c"

\item \code{wisc.secondary}: The six secondary colours of the University of Wisconsin's color palette

\item \code{wisc.accent}: The six accent colours of the University of Wisconsin's color palette

\item \code{wisc.heatmap.20}: Twenty tints of Twenty tints of the University of Wisconsin Badger Red to the lightest tint in wisc.badgerRedTints

\item \code{wisc.heatmap.100}: One-hundred tints of Twenty tints of the University of Wisconsin Badger Red to the lightest tint in wisc.badgerRedTints

}}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
\keyword{datasets}
