% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAnalysis.R
\name{word.count}
\alias{word.count}
\title{Word Count}
\usage{
word.count(comment)
}
\arguments{
\item{comment}{string of interest}
}
\value{
integer indicating the number of words in the provided string.
}
\description{
Count the number of words in a (text) string. A word is considered
any number of characters seperated by at least one physical space.
}
\examples{
 word.count("today is awesome!")
 # 3

 word.count("no")
 # 1

 word.count("")
 # 0

 word.count("yes, 3 is a word.")
 # 5

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
