% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean.DoB}
\alias{clean.DoB}
\title{Clean Date of Birth}
\usage{
clean.DoB(DoB)
}
\arguments{
\item{DoB}{string with date of birth-like data information.}
}
\value{
date with determined date of birth
}
\description{
This function uses the \code{\link[lubridate:as_date]{lubridate::as_date()}} function to determine
the date of birth by trying
\itemize{
\item Month - Day - Year
\item Year - Month - Day
\item Day - Month - Year
}
}
\examples{
clean.DoB("03301995")
# "1995-03-30"

clean.DoB("03/30/1995")
# "1995-03-30"

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
