% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.fromto}
\alias{convert.fromto}
\title{Phrase/Integer Conversion}
\usage{
convert.fromto(responses, fromto.tb, from, to)
}
\arguments{
\item{responses}{column to be converted; \emph{e.g.}, \code{Q1}}

\item{fromto.tb}{\code{tibble} (or \code{data.frame}) with the matched convertee and
converted pairs; see the example below.}

\item{from}{column in the \code{fromto.tb} containing the characters to be converted}

\item{to}{column in the \code{fromto.tb} containing the characters being converted to}
}
\value{
a vector of converted characters/phrases
}
\description{
Convert a collection of characters to another set of characters.
}
\details{
The function was initially designed to convert Likert responses to
integers, but it was quickly realized that it could easily be used for a
multitude of sins against data. The function relies on a user provided
\code{tibble} or \code{data.frame} with two columns; one with the characters to be
converted and the characters to be converted to.

The function is designed to work with \code{\link[dplyr:mutate]{dplyr::mutate()}} allowing multiple
conversions in a single command.

\emph{\strong{Note}}: Ideally, the integers are positive and non-zero.
}
\examples{
set.seed(13)
phrase2int.tb <- tibble::tibble(phrase=c("hated it!", "meh", "loved it!"),
                                integer=c(-1, 0, 1))
responses.words <- sample(x=c("hated it!", "meh", "loved it!"), size=5, replace=TRUE)
responses.integers <- sample(x=c(-1, 0, 1), size=5, replace=TRUE)

convert.fromto(responses=responses.words,
               fromto.tb=phrase2int.tb,
               from="phrase", to="integer")
# [1]  1 -1  0 -1  0

convert.fromto(responses=responses.integers,
               fromto.tb=phrase2int.tb,
               from="integer", to="phrase")
# [1] "meh"       "hated it!" "loved it!" "hated it!" "meh"

\dontrun{
mutate(tibble.oi, Q1.ints=convert.fromto(responses=Q1,
                                         fromto.tb=phrase2int.tb,
                                         from="phrase", to="integer"))
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
