% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_datasets.R
\name{dataset.summary}
\alias{dataset.summary}
\title{Dataset Summary}
\usage{
dataset.summary(
  dataset,
  ExcelFileName,
  n.examples = 4,
  overwriteXLS = FALSE,
  group.same.cols = TRUE
)
}
\arguments{
\item{dataset}{\code{tibble} or \code{data.frame} of interest}

\item{ExcelFileName}{string indicating the Excel workbook filename. The value
is passed to \code{\link[WriteXLS:WriteXLS]{WriteXLS::WriteXLS()}}. \emph{\strong{Re-running this command will overwrite
previous versions of the file!!}}}

\item{n.examples}{integer value indicating the number examples to return. Passed
to \code{size} of \code{\link[=extract.unique]{extract.unique()}}.}

\item{overwriteXLS}{logical to overwrite existing Excel workbook; default is \code{FALSE}}

\item{group.same.cols}{logical indicating if the columns should be grouped by
those with the same information.}
}
\value{
tibble of column names, types, and examples
}
\description{
When working with new datasets it is helpful to know what type
of data is included and easily retype the columns and provide new column
names. This function identifies the type of data in each column and provides
unique examples data within each column. The results are written to an Excel
workbook. The new column names and re-defined column types can easily be added
to the Excel workbook, imported into R, and assigned to the original dataset.
The goal is to reduce the logistical burden of the user.

This function will likely evolve overtime.

\emph{\strong{Re-running this command will overwrite previous versions of the file!!}}
}
\examples{
dataset.summary(dataset=ds.orig,
                ExcelFileName="ds_Column-names-and-data-types-and-examples.xlsx",
                n.examples=4)

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
