% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{LatLonDistances}
\alias{LatLonDistances}
\title{Latitude-Longitude Distances}
\usage{
LatLonDistances(lat1, lon1, lat2, lon2, distance.units = "miles")
}
\arguments{
\item{lat1}{latitude value for the first set of latitude-longitude coordinates}

\item{lon1}{longitude value for the first set of latitude-longitude coordinates
\emph{e.g.}, c(lat1, lon1), c(42.331177, -83.046023)  ## Detroit, Michigan}

\item{lat2}{latitude value for the second set of latitude-longitude coordinates}

\item{lon2}{longitude value for the second set of latitude-longitude coordinates
\emph{e.g.}, c(lat2, lon2), c(42.737652, -84.483788)  ## East Lansing, Michigan}

\item{distance.units}{string indicating if the resulting distance should be
miles or kilometers expressed as \code{"km"}; default is \code{"miles"}.}
}
\value{
distance between the two points taking into consideration the curvature of the earth
}
\description{
Calculate distance between points on the globe
}
\details{
long description with what the function does and returns
}
\examples{
point1 <- c(42.331177, -83.046023)  ## Detroit, Michigan
point2 <- c(42.737652, -84.483788)  ## East Lansing, Michigan

LatLonDistances(lat1=point1[1], lon1=point1[2],
                lat2=point2[1], lon2=point2[2], distance.units="miles")
# 78.48957

LatLonDistances(lat1=42.331177, lon1=-83.046023,
                lat2=42.737652, lon2=-84.483788, distance.units="miles")
# 78.48957

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
