% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_donut.R
\name{make.donut.data}
\alias{make.donut.data}
\title{Construct Donut Plot Data}
\usage{
make.donut.data(
  data,
  category,
  facetBy = NULL,
  layerBy = NULL,
  layer.order = "descend",
  layer.alpha.min = 0.75,
  category.order = "count",
  levels.rev = FALSE,
  category.count = NULL,
  r.inner = 4,
  r.outer = 6
)
}
\arguments{
\item{data}{\code{tibble} (or \code{data.frame}) with the column of interest. \emph{\strong{NOTE}}:
Do \strong{NOT} use counted data.}

\item{category}{string with column of interest containing the categories to
comprise the donut (\emph{aka} ring). Only provide \strong{ONE} column name.}

\item{facetBy}{string indicating the column to group data by; for when you
want to \strong{facet} your donut plots via \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}; see
\code{\link[=make.donut.plot]{make.donut.plot()}}.}

\item{layerBy}{string indicating the column to group data by; for when you
want to \strong{add layers} your donut plots; see \code{\link[=make.donut.plot]{make.donut.plot()}}.}

\item{layer.order}{string indicating the order of the layers. There are four
options:
\itemize{
\item \code{ascend} where the inner ring (donut) has the smallest value and the outer
donut has the greatest value
\item \code{descend} where the inner donut as the largest value and the outer ring
has the smallest value
\item \code{alphabetical} where the rings are ordered alphabetically starting from
the inner ring
\item \verb{user defined} where the user provides the order of the donuts starting
from inner ring. Only layers included in the vector (\emph{e.g.}, \code{c("4", "r", "f")})
are included in the resulting data.
}

For \code{alphabetical}, \code{ascend}, and \code{descend} only the first \emph{\strong{two}} characters
are needed.}

\item{layer.alpha.min}{value indicating the minimum alpha value; default: \code{0.75}}

\item{category.order}{string indicating if you want the data to be ordered by
\code{"count"} in \emph{\strong{decreasing}} order or by \code{"category"} in alphabetical order;
default: \code{"count"}.}

\item{levels.rev}{logical indicating if the order of the categories should be
reversed.}

\item{category.count}{string with the column containing "counts" for each "category."
This parameter is \emph{\strong{required}} when the count (or total) for each row was
pre-calculated and allows for the creation of donut data when raw data is
not available and one only has the summarised values.}

\item{r.inner}{numeric value defining the inner radius of the donut; default: \code{4}}

\item{r.outer}{numeric value defining the outer radius of the donut; default: \code{6}}
}
\value{
\code{tibble} with the data to construct the donut plot.
}
\description{
Construct the \code{data.frame} for the \code{\link[=make.donut.plot]{make.donut.plot()}} function.
The raw \code{data.frame} or a summarised \code{data.frame} is acceptable and will be
converted into the needed data for the construction of a donut plot.
}
\examples{
\dontrun{
  donut.DATA <- make.donut.data(data, category, facetBy=NULL,
                                category.order="count", levels.rev=FALSE,
                                r.inner=4, r.outer=6)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
