% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_donut.R
\name{make.donut.plot}
\alias{make.donut.plot}
\title{Construct Donut Plot}
\usage{
make.donut.plot(
  donut.DATA,
  colour.palette,
  label.col = NULL,
  label.size = 3.5,
  facet.nrow = NULL,
  facet.ncol = NULL
)
}
\arguments{
\item{donut.DATA}{\code{tibble} (or \code{data.frame}) constructed via \code{\link[=make.donut.data]{make.donut.data()}}}

\item{colour.palette}{vector of strings containing the colour palette.}

\item{label.col}{string with the column containing the labels. Setting to \code{NULL}
will result in no labels.}

\item{label.size}{label sizes; default: \code{3.5}}

\item{facet.nrow}{numeric \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} rows; default: \code{NULL}}

\item{facet.ncol}{numeric \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} columns; default: \code{NULL}}
}
\value{
ggplot2 graphic object
}
\description{
add the description of making the plot
}
\examples{
\dontrun{
  donut.DATA <- make.donut.data(data, col.oi, facetBy=NULL,
                                category.order="count", levels.rev=FALSE,
                                r.inner=4, r.outer=6)
  donut.PLOT <- make.donut.plot(donut.DATA, colour.palette=msu.palette,
                                label.col=NULL, label.size=3.5,
                                facet.nrow=NULL, facet.ncol=NULL)
}

}
\author{
Emilio Xavier Esposito \email{emilio@msu.edu}
(\url{https://github.com/emilioxavier})
}
